Added script for scraping wikipedia
This script will take the name of a wikipedia article and scrape the
article markdown and the markdown of the articles it links
to.

Since the article names are not always filesystem-friendly (contain
disallowed characters or characters that need to be escaped), the
article names are sanitized e.g. Python_(Programming_Language) becomes
Python_Programming_Language and will be saved as
Python_Programming_Language.md.

The wikilinks in the articles are also sanitized this way. This makes
importing into VoodooPad simpler because we can use the sanitized
article name as the page name.

This script requires the MediaWiki to Markdown converter available
here: https://github.com/philipashlock/mediawiki-to-markdown

The script uses wget to get the article, and then runs it through
the converter and then saves it to the filesystem.

example usage:

python3 scrape-wikipedia.py napoleon

Will download ~600 articles