Added script to import multiple files into a VoodooPad document
python3 voodoopad_import.py <document> <directory>

Where <document> is the VoodooPad document and <directory> is the
directory containing the files to import. The extension is removed
from the file to get the page name. E.g. if there was a file called
Regular_Expressions.md, the page name when importing into VoodooPad
will be Regular_Expressions.