Refactored encrypt/decrypt into VPEncryptionContext class
This class will handle the encrypting and decrypting of VoodooPad documents.

It has methods for reading/writing files. The encryption/decryption is
handled transparently. All that is required is the path to the VoodooPad
document and the password.

Currently it cannot convert an unencrypted document to an encrypted
one, or vice-versa. It can create new, or read and modify existing
encrypted documents.