Various changes to allow Voodoo pad to run inside Geekbench
Geekbench adds a layer between python and some of the filesystem
methods. pathlib does not play nice with this layer since it expects
a string instead of a Path object. Converting a Path object to string
before using any of the filesystem methods helps mitigate this.

Decryption has been disabled for now since it depends on an external
module. It can be re-enabled by importing vpenc in datastore.py.