Our current "render" code works like this:
for each keyword
  continually call find() on the text to locate all instances of the keyword and
  store the indices

  replace all instances of that keyword with the appropriate Markdown formatting

This needs to be a case-insensitive comparison so we call lower() on the text. The
problem is we were doing text.lower().find(). This means every time we called
find(), we were calling lower(). When running our Geekbench dataset, it was
spending 40% of the entire workload in lower().

This commit converts the text to lower outside of the find() loop. This brings
it down to only spending 17% of the time in lower(). This can probably be improved.